// src/frontend/react-admin/src/generated/models/user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BureauResponse } from './bureau-response';
import { ProfileResponse } from './profile-response';
import { RoleResponse } from './role-response';
/**
 * 
 * @export
 * @interface UserRequest
 */
export interface UserRequest {
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    nom?: string;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    prenom?: string;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    sexe?: string;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    email?: string;
    /**
     * 
     * @type {Date}
     * @memberof UserRequest
     */
    dateNais?: Date;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    tel?: string;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    poste?: string;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof UserRequest
     */
    password?: string;
    /**
     * 
     * @type {ProfileResponse}
     * @memberof UserRequest
     */
    profile?: ProfileResponse;
    /**
     * 
     * @type {BureauResponse}
     * @memberof UserRequest
     */
    bureau?: BureauResponse;
    /**
     * 
     * @type {Array<RoleResponse>}
     * @memberof UserRequest
     */
    roles?: Array<RoleResponse>;
}